+++ cephfs/cephfs.go
@@ -9,35 +9,10 @@ package cephfs
 import "C"
 
 import (
-	"fmt"
 	"unsafe"
 
-	"github.com/ceph/go-ceph/internal/errutil"
 	"github.com/ceph/go-ceph/rados"
 )
-
-// revive:disable:exported Temporarily live with stuttering
-
-// CephFSError represents an error condition returned from the CephFS APIs.
-type CephFSError int
-
-// revive:enable:exported
-
-// Error returns the error string for the CephFSError type.
-func (e CephFSError) Error() string {
-	errno, s := errutil.FormatErrno(int(e))
-	if s == "" {
-		return fmt.Sprintf("cephfs: ret=%d", errno)
-	}
-	return fmt.Sprintf("cephfs: ret=%d, %s", errno, s)
-}
-
-func getError(e C.int) error {
-	if e == 0 {
-		return nil
-	}
-	return CephFSError(e)
-}
 
 // MountInfo exports ceph's ceph_mount_info from libcephfs.cc
 type MountInfo struct {
